Update default hyperparameters
The purpose of this change is to improve performance of the model on default hyperparameter, particularly for TabNet.

Update to an exponential decay learn rate scheduler and Adam optimizer with learn_rate = 0.02, decay = 0.9 based on typical hyperparameters from the tabnet original paper https://arxiv.org/pdf/1908.07442.pdf, page 15.

This also increases default epochs to 100 in LearnerRegKeras and Learner TabNet and adds tensorflow as a dependency.